Added the LightActivity in order to leave the Main one for the menu
